requirements.txt
asttokens==2.4.1
backcall==0.2.0
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
dataclasses==0.6
debugpy==1.8.1
decorator==5.1.1
dominate==2.9.1
executing==2.0.1
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
future==1.0.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.34.0
importlib-metadata==7.0.1
importlib_resources==6.1.2
ipykernel==6.29.3
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.5
matplotlib==2.2.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
pooch @ file:///croot/pooch_1695850093751/work
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.4.1
pyzmq==25.1.2
requests @ file:///croot/requests_1707355572290/work
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
tifffile==2023.7.10
torch==2.2.1
torchvision==0.17.1
tornado==6.4
traitlets==5.14.1
triton==2.2.0
typing_extensions==4.10.0
urllib3 @ file:///croot/urllib3_1707770551213/work
wcwidth==0.2.13
zipp==3.17.0
